Move maven dependency resolving to infra package
