+++ sit-cv-core/src/main/java/io/sitoolkit/cv/core/domain/classdef/javaparser/DependencyFinder.java
@@ -1,0 +1,12 @@
+package io.sitoolkit.cv.core.domain.classdef.javaparser;
+
+import java.nio.file.Path;
+import java.util.Collection;
+
+public interface DependencyFinder {
+
+    boolean canFindDependencies(Path projectDir);
+
+    Collection<Path> findJarPaths(Path projectDir);
+
+}
